progress

+++ src/dhcomplyStandardLibrary.h
@@ -19,7 +19,7 @@ typedef unsigned char uint8_t;
 typedef unsigned short uint16_t;
 typedef unsigned int uint32_t;
 typedef unsigned long uint64_t;
-typedef unsigned long long uint128_t;
+typedef __uint128_t uint128_t;
 
 void valid_file_pointer(FILE *);
 void valid_memory_allocation(void *);
